requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
urllib3==2.2.3
